Moved serverManager files in both include and src to subdirectory server_manager.
