+++ src/server_manager/serverManagerArguments.c
@@ -4,8 +4,8 @@
 
 // MY LIBRARIES
 #include "ipTools.h"
-#include "serverManagerArguments.h"
-#include "serverManagerTools.h"
+#include "server_manager/serverManagerArguments.h"
+#include "server_manager/serverManagerTools.h"
 
 void print_server_manager_arguments(struct server_manager_arguments arguments)
 {
